hwmon: w83795: Fix type incompatibility with non-macro find_closest
The pwm_freq_cksel0 array was previously declared as u16 but used with
find_closest_descending(). Now that find_closest_descending() is
implemented as a function with 'int' parameters instead of a macro,
passing unsigned arrays causes type incompatibility errors. This patch
changes the array type from u16 to int to match the function's signature
and prevent compilation errors.

+++ drivers/hwmon/w83795.c
@@ -273,7 +273,7 @@ static inline s8 temp_to_reg(long val, s8 min, s8 max)
 	return clamp_val(val / 1000, min, max);
 }
 
-static const u16 pwm_freq_cksel0[16] = {
+static const int pwm_freq_cksel0[16] = {
 	1024, 512, 341, 256, 205, 171, 146, 128,
 	85, 64, 32, 16, 8, 4, 2, 1
 };
